// src/components/App/Home/Playlists/Settings/interfaces.tsx
import React from 'react'
import {ViewStyle, TextStyle} from "react-native"

// Rename

export interface I_RenamePlaylistProps {}

export interface I_RenamePlaylistStyles {
    container: ViewStyle;
    inputText: TextStyle;
    bar: ViewStyle;
    input: TextStyle;
    buttons: ViewStyle;
    cancel: ViewStyle;
    cancelText: TextStyle;
    update: ViewStyle;
    updateText: TextStyle;
}